requirements.txt
Pillow==8.0.1
pkg-resources==0.0.0
qrcode==6.1
six==1.15.0
